Fix the 'full_target'
Fix the 'full_target' variable name conflict while building sdk_addon. The 'full_target' name in the sdk_addon.mk file conflicts with the same one in the droiddoc.mk file, it will overwrite the variable , and broke the droiddoc building. Replace all the 'full_target' text with another, to avoid the conflict.